python_projects/requirements.txt
asttokens==3.0.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
clickhouse-connect==0.8.16
comm==0.2.2
cryptography==45.0.2
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
Faker==37.1.0
greenlet==3.1.1
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kafka==1.3.5
kafka-python==2.2.4
lz4==4.4.4
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pycparser==2.22
Pygments==2.19.1
PyNaCl==1.5.0
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==26.4.0
requests==2.32.3
scp==0.15.0
six==1.17.0
SQLAlchemy==2.0.40
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
zstandard==0.23.0
